Rename EpicNumericUpDown to Base1NumericUpDown
